Safari-609.1.11.1/LayoutTests/inspector/canvas/memory-expected.txt

Test that CanvasManager tracks canvas memory costs and is notified of changes.


== Running test suite: Canvas.memory
-- Running test case: Canvas.memory.memoryCost
Memory cost of canvas is NaN.

-- Running test case: Canvas.memory.canvasMemoryChanged
Change size of canvas to 200x200.
Memory cost of canvas updated to 166400.
